refactor: websocket transferred to api folder [WTEL-5297](https://webitel.atlassian.net/browse/WTEL-5297)
