+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@webitel/ui-sdk",
-  "version": "24.12.93",
+  "version": "24.12.94",
   "private": false,
   "scripts": {
     "dev": "vite",
@@ -70,9 +70,6 @@
     },
     "./src/validators*": {
       "import": "./src/validators*"
-    },
-    "./src/websocket*": {
-      "import": "./src/websocket*"
     },
     "./src/modules*": {
       "import": "./src/modules*"
